Add files via upload
This project demonstrates how to classify images using a dataset called CIFAR-10. It uses a type of artificial intelligence called convolutional neural networks (CNNs) to learn and recognize objects in images.
The goal is to train a model to identify 10 different types of objects, like planes, cars, and animals. 
The project covers loading the dataset, training the model, and using it to predict objects in new images.